separate out db helper functions
