+++ routes/locations.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router  = express.Router();
-const { getMarkers, getMarker, addMarker, editMarker, deleteMarker } = require('../maps-helper');
+const { getMarkers, getMarker, addMarker, editMarker, deleteMarker } = require('../marker-helper');
 
 
 // this will go on individual map page
@@ -23,13 +23,9 @@ router.post('/', (req, res) => {
 
 
   addMarker(userID, locations, db)
-    .then((res) => {
-      console.log("res and addmarker", res);
-    })
     res.redirect("/");
 })
 
 
 return router;
 };
-
